requirements.txt
Flask
Flask-SQLAlchemy
requests
apscheduler
python-dotenv